+++ seed_data/test_extract.py
@@ -1,4 +1,3 @@
-import asyncio
 from aioresponses import aioresponses
 import aiofiles
 import pytest
